+++ src/lightroom/__init__.py
@@ -32,7 +32,7 @@ from .exceptions import (
     PluginHandshakeError,
 )
 
-__version__ = "0.3.0"
+__version__ = "0.3.1"
 
 __all__ = [
     "LightroomClient",